Update diagram and separate TriggerRecordHeaderData into its own header file
